Use deploy_env instead of bosh-deployment
With the migration to paas-bootstrap we want to manage multiple
deployments from the same bosh. That prevents us from using
bosh-deployment tag to uniquely identify the environment.

We added instead the tag deploy_env which identifies the environment.

In this commit we stop using the bosh-deployment tag, and we use
deploy_env.

The metrics from the datadog agent will have the tag from the changes
done in [1] and the concourse ones from [2]

[1] https://github.com/alphagov/paas-cf/pull/696
[2] https://github.com/alphagov/paas-bootstrap/pull/8