Add DNS record for logsearch-ingestor
In [1] we moved the runtime-config from paas-cf to paas-bootstrap.

But one of the addons is syslog-forwarder, which requires a destination
syslog server to send the logs to.

Before we were using the ELB address from the terraform outputs, but
during the bootstrap from paas-bootstrap, this ELB is not generated.

To workaround this issue, we configure the syslog-forwarded to
a known generated DNS name, that is created as an alias of the ELB
endpoint when it is deployed.

[1] https://github.com/alphagov/paas-bootstrap/pull/34