Pass concourse password as pipeline parameter
The password is already known before the pipeline is pushed. We now pass
it as a parameter. This avoids requiring a resource to fetch it from the
bucket.

Also, the concourse password is not required anymore in the self-update-pipeline
task as self-update-pipeline runs environment.sh that already retrieves the
password from the bucket.