Read MFA token before writing token file.
Previously if a user cancelled (Ctrl-C) the read call asking for the MFA
token, the script would exit leaving a token file with just the EXPIRES
header, which means future runs of the script will think there's already
a valid token.

Re-ordering the script so that this header isn't written until after
reading the MFA token prevents this.