Add encrypted RDS broker plans
Add two new plans with Storage Encryption enabled[1]

As explained in the related ADR[2], in AWS:

 * Storage Encryption can only be enabled on creation of the DB.
 * Storage Encryption is only supported for some instance types.

In consequence:

  * Users **can not** migrate from a plan with encryption (Medium or Large)
    to a plan without encryption (Small or Free).
  * Users **can** migrate from Free/Small to a plan with encryption
    (Medium or Large), but the instance will not have encryption enabled.
  * The existing databases will remain without encryption.

We decided to provide additional explicit plans with Encryption enabled,
and keep the old ones.

We will add logic in the broker to prevent updates between plans with
or without encryption.

Add more plans is easy, but removing them is painful if they are being
used. We will limit this to the HA plans to reduce this risk.

[1] http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/Overview.Encryption
[2] https://government-paas-team-manual.readthedocs.io/en/latest/architecture_decision_records/ADR015-rds-storage-encryption-plans/