Update broker to not allow update encrypt settings
Update the rds-broker:

AWS RDS Storage Encryption[1] can only be set during creation of the
RDS instance. In consequence, it would be erronous and misleading
to allow update an instance to a plan that has different Storage
Encryption settings  (enabled/disabled or KMS Key ID).

We add logic to fail noisily to inform the user about the imposibility
of change encryption settings.

[1] http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/Overview.Encryption.html