Change user facing email address in user scripts
When we create a user or rotate their password, we want to send email from
the support address, so that when they respond, it goes to support and there
is someone to deal with it.

Note that for the scripts to be able to use this address, this has to be
verified in AWS SES. It's a manual 2-step process, where you have to receive
a mail and click on confirmation link. This has been done for all our
accounts.