Separate cloud_config in manifest tests.
We were previously combining the manifest and cloud-config into a single
YAML blob to be used in the tests. This does not reflect how these are
used in the system, so it seems better to separate these out, and be
explicit which we are referring to in the relevant tests.

I think these were initially combined like this to ease the transition
to cloud-config in the first place.