Reduce BBS max connections to database
The default is 200. But we use a t2.micro instance shared with CF and
UAA so we should make sure to not impact them negatively by opening too
many connections and exhausting the memory.

During our testing, BBS added 20 connections at most. This will be
sufficient for now but this number and the instance type should be
reviewed once we require more performance.