Remove override for CF RDS max_connections.
Production is currently using c. 56 connections to this RDS instance.
The default value of max_connections for a t2.small instance is 225
(2Gig / 9531392), which leaves plenty of headroom.

It's probably not a good ides to exceed the AWS default value for this
without thorough testing. If we do need nearer to 500 connections, we
should look to upgrade the instance to one with more memory.