Use terraform arrays for CIDR lists.
Now that we've upgraded to a version of terraform that supports array
variables, there's no longer a need to specify these as a comma
separated list. Using an array means we can split the entries onto
multiple lines which is much more readable (and allows us to add
comments for individual entries). It also saves having to split them in
the config.