Use constant-width formatting in availability test output.
This updates the output so that all the fields will align for most
reasonable cases. They'll be aligned for a run of up to 100k requests,
1k failed requests, or 10k seconds. For comparison, a recent 30min (1800
seconds) run generated 17k requests.

Unfortunately, there's no Format method for a time.Duration, so I've
resorted to converting to a floating-point number of seconds and
printing that to 2dp.

The updated formatting generated output like the following:

2017-01-20 15:56:14.12 UTC: Duration:  123.46s, Requests:  1234, Non 200 Requests:   0, Success Rate: 100.00%
2017-01-20 15:56:14.12 UTC: Duration:  123.46s, Requests: 12345, Non 200 Requests:   0, Success Rate: 100.00%
2017-01-20 09:04:06.12 UTC: Duration: 1739.40s, Requests: 23456, Non 200 Requests: 123, Success Rate:  99.54%