Use Concourse's get-version to pin cf-release
Currently we use the git-resource `tag_filter` to pin the cf-release
version.

But that is actually wrong, because `tag_filter` would only restrict the
versions that are discovered by the resource, not pin the version.
If, for instance, you checkout a branch with a different `tag_filter`
which would discover a newer version of the resource, our pipelines will
start using this latest version, even if you change back the
`tag_filter`.

The right approach would be define a `tag_filter` to discover all the
release tags (v* for cf-release) and pin the version using `get.version`
from concourse[1]

More info in this issue from concourse: [2]

[1] https://concourse.ci/get-step.html#get-version
[2] https://github.com/concourse/concourse/issues/256