Continuous smoke test failures monitor
We want to use the continuous smoke tests during out of hours support,
so we need to know when they are failing and be assured they are always
running.

To monitor smoke test failure:

We need to track the success/failure status of smoke tests and only
alert if consecutive tests fail, to rule out false alerts. Since
Concourse does not give us access to the historical build failure data
easily we use a custom metric on Datadog.

The monitor will attempt to notify us if the tests have been failing in
the past 3 runs. The notification, should go to the PagerDuty service,
which will take care of notifying appropriate support team member.

For assurance the smoke tests are always running:

A `no data` alert from a monitor which takes its metrics from the
continuous smoke tests job in Concourse will provide an indication the
smoke tests are no longer running. However, keep in mind, it could just
be symptomatic of the metrics not being sent or received properly.

We add a conditional notification to PagerDuty if there is no data
received for 15 minutes (Note the @pagerduty recipient needs to be set
up in the Datadog PagerDuty integration to receive the alerts).