Cleanup FIXMEs from CF RDS upgrade.
The parameter group had to be left in place from the upgrade (85bd51f)
because the upgrade wasn't applied immediately, instead it was applied
in the defined maintenance window for the instance.

Now that the upgrade has been applied, the old parameter group can be
removed.

Similarly, we no longer need to set allow_major_version_upgrade because
the upgrade has completed.