Pin cf-release with paas-cf from cf-deploy
Note: this commit is aligned with the change to pin the `get`
of cf-release in the commit 3f623b13f557df1df6e6e35c90a1684d8e3e91cf

Although it is not directly being used by the cf-deploy job, we pin the
version of cf-release using the `get.version` syntax[1] in that job.
We must define it as a input to prevent the pipechecker from failing:

```
==  paas-cf/concourse/pipelines/create-cloudfoundry.yml  ==
WARNING * Unused Fetch: job='cf-deploy', resource='cf-release'
make: *** [lint_concourse] Error 20
```

With this commit we maintain the consistency of the deployed paas-cf with
the deployed version of cf-release using the `passed` tag based on the
version pulled during the cf-deploy task.

In the case of the fail-pipeline, we pull the cf-release toguether with
paas-cf.

[1] https://concourse.ci/get-step.html#get-version