Remove monitor for garden debug endpoint
The garden debug endpoint added in [1] does not work in CF release v250,
because garden job does not configure the debug listen address by
default anymore[2].

We could enable the debug port, but after discussing it we consider that
there is not too much value on monitoring this port, it be a security
risk as we are exposing an additional port and we are already monitoring
if the process anyway.

We remove this monitor to avoid it from failing.

We should remove the check from the VM as well in the
paas-datadog-for-cloudfoundry.

[1] https://github.com/alphagov/paas-datadog-for-cloudfoundry-boshrelease/commit/1be63198ccc30ccbd38967588ac50e56d8121082
[2] https://github.com/cloudfoundry/garden-runc-release/commit/e22d9bf5143d9b9dbd3c2d4abd68d9c52a64d999