Update DD monitor options
Since we have discovered that go-datadog library has problems with
setting default options for some monitor
settings(https://github.com/zorkian/go-datadog-api/issues/56) we need to
use this script to ensure all options that we care about have desired
values.

Also, as ruby library for DataDog api is re-creating a
monitor every time it does update we need to send all options that were
previously set for the monitor to not fallback to monitor defaults.