Stop using separate etcd release.
We initially used a separate etcd release from the one bundled in
cf-release because diego recommended a different version to cf-release
(1c57c51). Now that we've migrated diego to use the relational backend,
this is no longer a concern.

CF v251 includes etcd version 87, so this change is effectively a no-op.