Reduce BBS max_open_connections to 100.
Now that the deadlock issue is addressed, there's no need for it to be
this high. Additionally with the updated CF release we've been hitting
DB errors which look to be caused by us hitting the Postgres
max_connections limit. Reducing this back down to 100 will leave a lot
more headroom which should avoid this problem.