Add monitors for diego services locks
We add monitors for several services from diego: bbs, nsync,
tps and auctioneer, in the same way than we did in the route_emitter in
531b9e92659f490ccdd91c76904b5dcb34159202

The services report metrics cf.<service>.LockHeld.v1-locks-<service>_lock
which  measures whether a given holds the route-emitter lock: 1 means
the lock is held, and 0 means the lock was lost.
The metric is emitted every 30 seconds by the process which holds the
lock.

As there can be only one nstance of the service holding the lock,
this metric can be only 1 for each environment.
Checking this value would allow us detect if there are
two services hold the same lock thinking that are the master,
as it happened before with route emitter[1].

The monitor must be exactly one, so we want to check upper and lower
thresholds. But Datadog does not allow that kind of check, and 2
different monitors must be created. We use `count=2` to create bosh and
we use `element()` and `format()` to build the right queries and
thresholds.

[1] https://github.com/cloudfoundry/diego-release/issues/201