Add monitor for etcd unique leader presence
We added monitors for several services from diego: bbs, nsync,
tps, route_emitter and auctioneer in previous commits to check that
there is only instance of the service as master.

In this commit we add a similar monitor checking that there is at least
and only one etcd leader in the deployment.

For that we check the metric `cf.etcd.IsLeader`, which is sent by
the service from the bosh etcd-release `etcd_metrics_server`[1][2].
1 means the node is leader, and 0 means not.
The metric is emitted every 60 seconds by default.

As there can be only one instance of the service being leader,
this metric can sum only 1 for each environment.
Checking this value would allow us detect if there are
two etcd repoting being leaders.

The monitor must be exactly one, so we want to check upper and lower
thresholds. But Datadog does not allow that kind of check, and 2
different monitors must be created. We use `count=2` to create bosh and
we use `element()` and `format()` to build the right queries and
thresholds.

[1] https://github.com/cloudfoundry-incubator/etcd-release/tree/master/jobs/etcd_metrics_server
[2] https://github.com/cloudfoundry-incubator/etcd-metrics-server/blob/ddc722acd43af2acc829614bd5197de6488956ea/instruments/server.go#L57