Pin cf-release version in tag_filter.
Due to the way the git resource is implemented[1], a newly pushed pipeline
with a wildcard tag_filter will only see the latest tag that matches the
glob. This means thay with the pipeline defined as it currently is, it
will only work from clean when we are deploying the altest version of
CF. As soon as a newer version is released, that's the only one that
will be returned, and the later version pin on the get step will never
find the required version, and therefore stall the pipeline.

[1]https://github.com/concourse/git-resource/blob/30851a3/assets/check#L76