Fix name of java buildpack.
The package was renamed in CF v246[1]. This omission wasn't an obvious
problem for existing deployments because the buildpack had already been
uploaded, however for a clean deployment the java buildpack would be
missing. For existing deployments it means that the buildpack wouldn't
have been updated from the previous version.

[1]https://github.com/cloudfoundry/cf-release/commit/72222a8