Add ability to create CDNs programmatically
We would like to create a CDN for the PaaS Documentation and the Product
Page and be able to access it, from the new subdomain.

The aim of any tool, is to make it easy to use. In this case as extensible
as possible, without making too many changes. Therefore to achieve that,
we are creating each distribution instance in form of a terraform
module. These modules are all stored in one `instances.tf` file.

The CNAME record, will be created as part of the system domain name.