Adjust UAA ELB timeout
From time to time requests fail on the UAA ELB with a 504 error. In most
cases this was caused by an unhealthy backend, but some cases were not
explained. This was investigated in #123760561 and AWS support ticket
1861267451 was raised.
The problem may happen when ELB maintains connections to the backend
because the backend uses HTTP keepalive, and the idle timeout on the ELB
is higher than the backend keepalive timeout. This is explained in:
https://hardwarehacks.org/blogs/devops/2015/12/29/1451416941200.html

We have reasonable understanding of what is going on on UAA: the ELB
backend here is tomcat, configured with a 20s keepalive timeout. The
risk is probably low on UAA as requests should never take a long time,
we change the ELB idle timeout to 19s in hope the ELB terminates the
connection instead of the backend.

This will be monitored over a period of time to check if things improve.