Remind operator to add user to the mailing list
As a part of improve incident communications story we've decided to
notify operator that every user should be a member of GOV.UK PaaS
Announce. As a first step before implementing automation for this task
we will would like to add notification that will be displayed after user
account was successfully created.