Simplify the threshold of route-emiter
In the route_emitter lock monitor, both the upper and lower limits are
set to 1, with the same value. Because that we do not need to have logic
to chose a value depending if we are configuring the upper or lower
threshold.

But the Datadog API fails obscurely when the threshold is different
than the value set in the query. This can get tricky if somebody wants
to change or reuse this pattern and define a different value for the
lower/upper limits. We add a comment to highlight this fact.