Use max/min for the upper/lower lock thresholds
As discussed in [1]

When we are querying the lock metric, use min or max in the last 5m
depending if we are checking the upper or lower threshold.

This would make the monitor more sensible to deployment for the upper
limit (lower was already min) but it would make it work again.

[1] https://github.com/alphagov/paas-cf/pull/746#issuecomment-277227901