Revert to default idle timeout on CC ELB
This was originally set on the router ELB and copied ever since to all
ELBs. See:
https://github.com/gds-attic/paas-alpha-cf-terraform/commit/4a7191d4330a59e7676c8ae3c449f0b2cfd84c3e

This might be the cause of 504 errors we see from time to time on the CC
ELB. Since the CC Nginx has a keepalive_timeout set to 75s, the ELB
should have a lower idle timeout. Reverting to default will set it to
60s.

See commit 51f443f2c3a9828687d7ade0693eb7a3ebea72ca for details.