Travis: set +e after each before_install
We added `set -e` before each `before_install` block in c1a5aa0 to make the
build fail fast if it was unable to install any of the dependencies.

There appears to have been a change to Travis today that now causes the
build to fail if `set -e` is used in `before_install`. From some previous
issues it looks like it causes it to be set globally for all of Travis's own
scripts that run within the build and has never been fully supported:

- https://github.com/travis-ci/travis-ci/issues/3586#issuecomment-274585174
- https://github.com/travis-ci/travis-ci/issues/891#issuecomment-32318632

We would still like the build to fail fast if it's unable to install the
dependencies. So it seems that the compromise is to unset it after each
command block until the following feature is implemented:

- https://github.com/travis-ci/travis-ci/issues/1066