Change parser VM type to medium
Due to huge volumes of logs to parse we went out of cpu credits on
parser VMs. This slowed down logstash considerably and resulted in
almost no log ingestion to elasticsearch.  To avoid such situation in
the future we would like to use a VM type that is not burstable.