Allow the pipeline to fail
Adding the `path` attribute, to the 'terraform/cf-certs/upload.tf' is
causing the task to fail. This is expected behaviour, as it redefines
the certificate.

Terraform would be trying to remove the old certificates, causing amazon
to fail, as these are still in use. The idea is, that the new
certificates will get created and defined to be used. Once that happens,
the old ones, will be peacefully removed.

This problem, will occur only in our DEV and CI environments. Staging
and prod should not be affected.

This commit, could be reverted after a week after merging this PR.