Reduce the default TTL for CloudFront instances
We have decided, it may be unreasonable to have a 60 minute delay,
before CloudFront is going to check for the changes.

This has been justified with the scenario: During user testing lab, we
may want to quickly run an update on the docs and expect it to appear
fairly quickly online, for the users to look up.

This has been changed to 5s as agreed with Ben, Dan, Alex and Michal. It
may be worth, keeping an eye on this change.