Adjust smoke tests notifications
Create explicit variables for alerting in and out of hours. They can be
set differently for each environment:

* empty string for no notification
* email address
* pagerduty service name

We currently set email notification for dev, in hours pagerduty only in
ci and staging, in hours and out of hours pagerduty in production.

Note: the "no data" smoke test monitor is moved to "in hours" alerting. We
want to check how frequent it is before moving it to "24x7"