Pingdom: add environment specific contacts for PD
In dev, we only want to send notifications to default email contact
(currently pointing to
the-multi-cloud-paas-team@digital.cabinet-office.gov.uk ).
In Staging and CI, we want to additionally notify in-hours Pagerduty.
In prod, we want to send notification to 24/7 pagerduty.

The contact IDs correspond to contact sending emails to the specific
PagerDuty integration endpoints.