TF lint: get pingdom provider yourself
Do a download into temporary dir, so that it can be used for lint.
This means we don't need to explicitly download it in travis.

The provider is required for `terraform graph` command in the TF 0.8.5
version. Otherwise TF will fail complaining that it can't find the provider
and lint won't work.