Make RDS Broker use SSL
In order for our Trusted Advisor to report when we desire, we would like
to fix some issues we're currently facing.

One of the issues, is that the RDS Broker could use HTTPS connection.

In order to prevent downtime, we're adding an extra listener, which
then will be removed in another PR.