Make syslog use SSL
In order for our Trusted Advisor to report when we desire, we would like
to fix some issues we're currently facing.

One of the issues is that the logsearch-ingestor could use ssl connection.

In order to prevent downtime, we're adding an extra listener, which
then will be removed in another PR.

Also, we need additional ingestor-elb security group here to allow TLS
traffic to newly created listener. It wasn't possible to use existing
group as during terraform apply it tried to delete it together with ELB
which wasn't possible as this group is also attached to ingestor VM's.

Terraform error:

```
Error applying plan:

1 error(s) occurred:

* aws_security_group.logsearch_ingestor_elb: DependencyViolation: resource sg-e0dce486 has a dependent object
	status code: 400, request id: 8bd5c2b2-db40-44e2-9682-ba25f42d4a17
```