Enable route-services
This will enable tenants to add an authentication layer in front of an
app without having to modify the app itself.

The CF documentation[1] says that the route_services_secret needs to be
a "robust passphrase", but don't give any more detail on what they mean
by this. From looking at the code[2], the passphrase specified in the
manifest is used to generate a 16 byte encryption key using the PBKDF2
key derivation function[3]. This is then used to construct an AWS-128
cipher (a 16 byte key selects this cipher[4]). Based on this assessment,
we believe that a "robust passphrase" in this context means one that
can't be easily brute-forced, and that this requirement is consistent
with most of the other passwords in use in our systems. We therefore
consider that the `simple` scheme is suitable for this one.

[1]https://docs.cloudfoundry.org/services/route-services.html#enabling-route-services-in-cloudfoundry
[2]https://github.com/cloudfoundry/gorouter/blob/d94cde9/main.go#L147-L155
[3]https://github.com/cloudfoundry/gorouter/blob/d94cde9/main.go#L147-L155
[4]https://golang.org/pkg/crypto/aes/#NewCipher