Use blue-green-deploy plugin for healthcheck app
This will ensure that there is no downtime when we deploy the healthcheck
app - which happens during every run of the pipeline. It will enable us to
merge the `post-deploy` job into `cf-deploy` because it will now be safe to
run in parallel with `availability-tests`.

Although it hasn't been a problem to date, it will also prevent us from
getting false-positive alerts from Pingdom (which monitors this app) and may
allow us to reduce the threshold for alerts.

For non-development environments, where `DISABLE_HEALTHCHECK_DB` is not set,
we can no longer bind the database service after pushing the app because it
would require a restart which would cause downtime. To workaround this I am
appending the service to the manifest within the conditional, so that the
service is bound while the app is pushed and started by the plugin.

I've used some inline Ruby because it was the easiest and safest way to
modify the manifest and it didn't seem like we could extract any generic
functionality to a script. I did consider adding the service to the checked
in manifest, but I don't think it's appropriate because the app is used with
a different service name in the custom acceptance tests and I frequently use
this app without a service for manual testing.