Delete healthcheck-old app after blue-green-deploy
The blue-green-deploy plugin leaves the `-old` suffixed application running
after a successful deployment. I can't see a reason for doing so, so I'm
adding a command to remove it if the previous deploy command has passed.

I've also raised a feature proposal about changing the behaviour of the
plugin in the future: bluemixgaragelondon/cf-blue-green-deploy#29