Create user: purge creator from org
There is an issue [1] where even after removing all the user roles from
spaces and orgs, the user is still present in the list of all users of the
org. Fix for this is to remove user explicitly from the org using v2 API.
Enhance user creation script to do this. Otherwise user who created org/
added new user will be left over as member of that org.

[1] https://github.com/cloudfoundry/cli/issues/781