Re-order post-deploy tasks to fix initial deploy.
The deploy healthcheck job was running in parallel with all the other
jobs. It requires the rds broker to be setup, along with the
corresponding application security groups. It also requires the
test_apps quota to be setup.

To avoid issues here, this therefore moves these required jobs earlier
in the process to run in parallel with the create-admin-org-space task.