Add redirection to add www to the system domain
We use a configuration based on Cloudfront + API Gateway to redirect to
the www.${apex} domain from the APEX system domain.

For that we create a module that can be reused to implement any
redirection and reused later, and we add one instance in the
terraform/cloudfront/instances.tf

The implementation configures an API Gateway that will redirect a 301 to
a new location for any GET or HEAD to the root path. We use cloudfront
to service this endpoint publically with SSL.

Limitation: Any other path will return a 403 error. In order to make
API Gateway work with any path, we must use proxy mode and send the
request to a Lambda function or similar.