Make Terraform version check more robust
This awk monstrosity (that I introduced in 50bb83f) was not working with
gawk 3.1.8 or mawk 1.3.3 (as available on Ubuntu Precise). It turns out
that the regex interval expressions (`{1,4}` etc) only work with gawk 3
if you specify the --posix flag. They're enabled by default with gawk
4+.

Instead of attempting to resolve this in a portable way, we can use the
new feature in Terraform[1] to specify constraints on the version
required. This was introduced in Terraform 0.8.0. Versions earlier than
this will error with "* Unknown root level key: terraform"

I considered adding the version constraint to globals.tf, but that's not
used by all our configs, so I thought it better to introduce a new file
that can be symlinked into every config dir.

[1]https://github.com/hashicorp/terraform/pull/10080