Allow our own build Concourse to push apps
We are going to use the build Concourse in CI to automatically push
apps, such as the tech docs, to our production platform. We need
to allow it to communicate with the API until the API is opened up
to the public. The IP is an Elastic IP, so it will be stable over
time.