Prepare for aws credentials validation
We want to have different behavior of AWS credentials check based on if the
script runs inside concourse pipeline, which is a server that uses IAM role
and when it runs on the command line in your laptops, in which case we use
STS tokens + temporary credentials. In order to introduce this logic, we
need to first add variable in self-update-pipelines where we store the
information for the script to not do STS related credentials check, but skip
them instead.