Don't rotate RDS passwords
RDS passwords take a while to rotate - depending on the settings of your
instance, this can be during the next maintenance window. This would bring
the password out of sync with configuration of the rest of the system, which
expects PW to be updated.

In case of CF, that would break BBS, which is core component. Don't rotate
RDS passwords. We might investigate this in the future and find a way to do
interruption-less rotation.