Remove *_monitors tag from CC monitors
They are constantly warning on the dashboard because of datadog showing
them in “No data” state. A support case is ongoing with datadog.

In the meantime, we remove the tag so the dashboard won’t be able to
find them. Since we have set up pagerduty alerting, we’re covered.

This should be reverted once the datadog bug is solved.