Revert "Increase number of CC workers to 3 per VM"
This reverts commit c292aad01133c3cb79117e1aca4bbd11e22bcf1d.

The workers are constantly polling the RDS CC DB with query:
SELECT * FROM "delayed_jobs" WHERE (((("run_at" <= '2017-03-06
19:00:02.044862+0000') AND ("locked_at" IS NULL)) OR ("locked_by" =
'cc_global_worker.api_worker.0.2')) AND ("failed_at" IS NULL) AND
("queue" IN ('cc-generic'))) ORDER BY "priority" ASC, "run_at" ASC LIMIT
5

This causes significant CPU usage and we have run out of credits. The
impact is that services connecting to the databases in that instance are
slower (api, uaa, bbs). This has already been reported by some users.

We now revert to the default. This is safe as this commit hadn't made an
improvement.