Increase CF DB instance type to db.m3.medium
We ran out of CPU credits in the CF RDS instance because of unexpectedly
high CPU usage. This caused CPU usage to be capped at 20% and it made
the system slow. We noticed most CPU was consumed by cloud controller
workers. They constantly poll the DB for new jobs to execute. We reduced
the number by reverting [1]

We change the instance type to a type that is not credit based.

[1]  https://github.com/alphagov/paas-cf/pull/817