Increase prod cell count to 12.
We're currently exceeding the limit defined in ADR017[1]. With 9 cells,
the total memory provided is 269G. The amount we need depends on
whether we include the Smoke tests and acceptance tests. The numbers are
as follows:

- Not including smoke tests: 268G
- Including smoke tests: 276G
- Including smoke tests and custom acc tests: 291G

Given the smoke tests are running continually, this usage should be
included in the total, hence the need to increase the cell count.

[1]https://government-paas-team-manual.readthedocs.io/en/latest/architecture_decision_records/ADR017-cell-capacity-assignment/