Omit silenced monitors from DataDog dashboard
I'm using the "mute for N hours" button in DataDog to silence some alerts
that we know are currently false-positives. I'd like them to be omitted from
the dashboard so that I can concentrate on real alerts.

I had to look at the API docs[0] to implement this. It seems that
`options.silenced` will contain a hash of `scope: timestamp` when a monitor
is muted/silenced. I'm omitting the monitor if the hash contains *anything*
because:

- I don't think the scope matters because the dashboard currently collapses
  triggered monitors of the same type, e.g. it tells us we have one warning
  if two hosts have alerts for disk space.
- I don't think we need to compare the timestamp because the `silenced` will
  be removed from the API results when the mute period expires.

[0]: http://docs.datadoghq.com/api/#monitor-get-all