Add script to generate buildpack changes notes
When upgrading CF, we often upgrade the buildpacks. The changes in the
buildpacks affect directly our tenants, and we must inform them about
the changes.

But gathering all the release notes is painful and slow.

This script automates the process, checking the buildpack versions added
between versions of cf-release, and automatically downloading and
aggregating the buildpack release notes. It writes them down in
a markdown file.

This file can be later sent as a update to the tenants.