dashboard: Separate no data alerts from warnings
A "no data" alert isn't necessarily as serious as a "warning". They normally
indicate that the monitoring isn't working correctly, which may be hiding
more serious problems, but aren't the same as a known alert.

I choose "unknown" as the dashboard label because "no data" could possibly
be confused with the dashboard no updating from the DataDog API.

This is particularly of interest to me at the moment while I'm on support
because we're investigating a problem with a handful of monitors that are
reporting "no data". I have been muting the affected monitors and dc41a8e
prevents them from showing on the dashboard, but because they come and go I
haven't been able to mute all of them, and I'm still getting distracted from
real alerts.