Split Consul client and server checks
Currently we run the Consul server checks on all VMs, because the
datadog-consul job in the datadog-for-cloudfoundry release couples
the two together. This splits them into those for the client and
those for the server. We leave the client checks on all VMs, but
only include the server check for the VMs which are part of the Consul
cluster.

To implement this, we needed to split jobs in
datadog-for-cloudfoundry-release first:
https://github.com/alphagov/paas-datadog-for-cloudfoundry-boshrelease/pull/18