Make the UAA ELB use sticky sessions
We began using passcodes for logging into the CLI using Google via OpenID
Connect. The user visits login.domain/passcode to get a one-time code to use in
the CLI, they then get redirected to /login, visit google to login and then
return to access the passcode page.

This requires sticky sessions to be enabled on the UAA load balancer because
the the URL that a user was intending to visit is saved in a server side
session and there does not seem to be a way of sharing the sessions between UAA
instances.

The behaviour with 2 UAA instances is that logging in consistently failed; we
think this may be because if at any point in the flow you hit another UAA
instance, you'll get a new JSESSIONID cookie and lose your old session
containing login state.

The default deployment of CF uses gorouter to load balance UAA requests.
gorouter automatically uses sticky load balancing whenever a JSESSION cookie is
encountered, as described here: https://docs.cloudfoundry.org/concepts/http-routing.html

In alphagov/paas-cf#196 we switched to using ELB to load balance UAA, in
order to allow us to implement an IP ACL/whitelist against UAA (and the other
API endpoints).

In order to fix the currently broken admin accounts, we're using ELB's
application-cookie based stickiness and basing it on the JSESSIONID cookie used
by UAA. The alternative time-based mechanism would probably work , but we
didn't use it because it seemed preferable to ensure the ELB cookie and the UAA
cookie lifetimes were automatically the same.